:recycle: refactor: rename Class name
